third_party/blink/web_tests/fast/forms/time/time-picker-select-invalid-min-value.html

<!DOCTYPE html>
<meta name=fuzzy content="maxDifference=0-3; totalPixels=0-1000">
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../resources/common.js"></script>
<script src="../resources/picker-common.js"></script>
<script src="../calendar-picker/resources/calendar-picker-common.js"></script>

<input type="time" id="time" value="14:15" min="10:12">
<script>

let t = async_test('Test select invalid min value in time popup');

function selectValue() {
  let timeElement = document.getElementById("time");
  timeElement.addEventListener("change", t.step_func_done(() => {
    assert_equals(timeElement.value, "06:16");
    assert_true(timeElement.validity.rangeUnderflow);
    assert_false(timeElement.validity.valid);
  }));
  clickTimeCellAt(0, 4);
  clickTimeCellAt(1, 1);
  clickTimeCellAt(2, 1);
  eventSender.keyDown('Enter');
}

t.step(() => {
  openPickerDeprecatedJsTest(document.getElementById('time'), t.step_func(selectValue));
});
</script>
